+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changes
 ### Additions
 - Json module implemented for config data structure
+- Added data folder with two samples and a config file for testing
 
 ## [0.1.8] - 2018-01-10
 ### Changes