+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Json module implemented for config data structure
 - Added data folder with two samples and a config file for testing
 - Added 2D transformation option to mp-transformation.py
+- Placing mp-transformation script in the cranium directory to function as a module
 
 ## [0.1.8] - 2018-01-10
 ### Changes