+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [Unreleased]
+### Changes
+### Additions
+- Json module implemented for config data structure
+- Added data folder with two samples and a config file for testing
+- Added 2D transformation option to mp-transformation.py
+- Placing mp-transformation script in the cranium directory to function as a module
+
 ## [0.1.8] - 2018-01-10
 ### Changes
 - Update matplotlib requirement from 1.5 to 2.0 to avoid installation problems with matplotlib dependencies for freetype and pngg