+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [Unreleased]
-### Changes
+### Changes 
 - Changed `convert_to_arr` to accomodate a main array and a list of additional arrays
 ### Additions
 - Json module implemented for config data structure